.github/workflows: maintenance scripts re: erigon-lib
erigon-lib is a (new) dependency that changes
a lot, and which requires custom patches
for Ethereum Classic support.

The automated CI script here used to fail because
of merge conflicts often; often these conflicts
were related to changes in the erigon-lib dependency.

So this patch installs a first attempt at
automating the maintenance of the etccooperative/erigon-lib fork
as a replaced go module dependency.

Date: 2023-02-14 08:34:43-08:00
Signed-off-by: meows <b5c6@protonmail.com>