.github/workflows: problem: cannot push to etccooperative/erigon because workflow perms
See https://github.com/ETCCooperative/erigon/actions/runs/4176375681/jobs/7232581868

Complaint of 'wont push to origin' because
token does not have 'workflows' permission.
But I just edited that env var to use a
newly-generated token that does, in fact, have
the workflows permission.

So now I've changing the token var to a newly
created repository secret; hoping that the problem
was secretly-cached secrets, and that using a new
variable avoids the cache false-positive on a token
value.

Date: 2023-02-14 09:35:00-08:00
Signed-off-by: meows <b5c6@protonmail.com>